requirements.txt
Flask
Flask-SQLAlchemy
requests
celery
